refactor: Refactor folder name
