+++ lib/src/commands/create_project.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:args/args.dart';
 import 'package:mineral/api.dart';
 import 'package:mineral/console.dart';
-import 'package:mineral/src/internal/entities/cli_manager.dart';
+import 'package:mineral/src/internal/managers/cli_manager.dart';
 import 'package:path/path.dart';
 
 class CreateProject extends MineralCliCommand {
