+++ lib/src/internal/managers/cli_manager.dart
