+++ lib/src/internal/managers/command_manager.dart
@@ -2,7 +2,7 @@ import 'dart:mirrors';
 
 import 'package:mineral/api.dart';
 import 'package:mineral/core.dart';
-import 'package:mineral/src/internal/entities/store_manager.dart';
+import 'package:mineral/src/internal/managers/store_manager.dart';
 
 class CommandManager {
   final Map<String, SlashCommand> _commands = {};
