+++ lib/src/internal/managers/reporter_manager.dart
