+++ lib/src/internal/managers/store_manager.dart
