+++ lib/src/internal/websockets/packets/guild_scheduled_event_delete.dart
@@ -1,8 +1,6 @@
-import 'dart:convert';
-
 import 'package:mineral/api.dart';
 import 'package:mineral/core.dart';
-import 'package:mineral/src/internal/entities/event_manager.dart';
+import 'package:mineral/src/internal/managers/event_manager.dart';
 import 'package:mineral/src/internal/websockets/websocket_packet.dart';
 import 'package:mineral/src/internal/websockets/websocket_response.dart';
