+++ lib/src/internal/websockets/packets/resumed.dart
@@ -1,5 +1,4 @@
 import 'package:mineral/core.dart';
-import 'package:mineral/src/internal/entities/event_manager.dart';
 import 'package:mineral/src/internal/websockets/websocket_packet.dart';
 import 'package:mineral/src/internal/websockets/websocket_response.dart';
